+++ datahub/search/contact/test/test_elasticsearch.py
@@ -118,6 +118,16 @@ def test_mapping(setup_es):
                             'copy_to': ['company.name_trigram'],
                             'normalizer': 'lowercase_asciifolding_normalizer',
                             'type': 'keyword',
+                            'fields': {
+                                'keyword': {
+                                    'normalizer': 'lowercase_asciifolding_normalizer',
+                                    'type': 'keyword',
+                                },
+                                'trigram': {
+                                    'analyzer': 'trigram_analyzer',
+                                    'type': 'text',
+                                },
+                            },
                         },
                         'name_trigram': {
                             'analyzer': 'trigram_analyzer',
@@ -126,6 +136,12 @@ def test_mapping(setup_es):
                         'trading_names': {
                             'copy_to': ['company.trading_names_trigram'],
                             'type': 'text',
+                            'fields': {
+                                'trigram': {
+                                    'analyzer': 'trigram_analyzer',
+                                    'type': 'text',
+                                },
+                            },
                         },
                         'trading_names_trigram': {
                             'analyzer': 'trigram_analyzer',